INTEGER_LLM v0.91.0 (theta_v 0.22.0, kernels 0.65.0, runtime 0.62.0), CLIENT v0.59.0 (myl-client   0.44.0): das hybride 35B antwortet, und die Ursache stand in der ersten Stufe der ersten Ebene; Funde   417 bis 430 | theta_v/spec.json (silu_zerlegt, drei Festlegungen der Zustandsschicht);   INTEGER_LLM/kernels (integer_math, faltung, zustandsschicht, rmsnorm); INTEGER_LLM/runtime (model.rs,   loader.rs, bin/zustandsprobe, bin/logit_probe); INTEGER_LLM/calibrate (quantize.py, model_configs.py,   main.py); artifacts (alle fuenf neu gebaut), scale_packs (alle fuenf neu erzeugt);   MODELS/llm/KATALOG.json (35B ergaenzt); CLIENT/myl-client (oertlich.rs, tests/chatvorlage.rs); beide   Fahrplaene, Master v1.281.0, sauberkeitsprobe.py

+++ INTEGER_LLM/kernels/src/zustandsschicht.rs
@@ -439,7 +439,7 @@ pub fn schritt(
 
     // Die Verschiebung, die den Groesstwert gerade noch in i16 legt.
     let groesst = akkus.iter().map(|a| a.unsigned_abs()).max().unwrap_or(0);
-    let bits = 128 - u32::from(groesst.leading_zeros());
+    let bits = 128 - groesst.leading_zeros();
     // ⚠️ Nach oben begrenzt, damit `aus_frac` nicht unter null faellt;
     //    darueber bleibt es bei der ausdruecklichen Saettigung.
     let mut schiebung = bits.saturating_sub(15).min(ZUSTAND_FRAC + NORM_FRAC);
